Separate state timeline tooltip component, rely on field display color to retrieve color of series
