+++ public/app/plugins/panel/state-timeline/StateTimelinePanel.tsx
@@ -1,18 +1,10 @@
 import React, { useCallback, useMemo } from 'react';
-import {
-  DataFrame,
-  FALLBACK_COLOR,
-  formattedValueToString,
-  getDisplayProcessor,
-  getFieldDisplayName,
-  getValueFormat,
-  PanelProps,
-} from '@grafana/data';
+import { DataFrame, PanelProps } from '@grafana/data';
 import { TooltipPlugin, useTheme2, ZoomPlugin } from '@grafana/ui';
 import { TimelineMode, TimelineOptions } from './types';
 import { TimelineChart } from './TimelineChart';
-import { findNextStateIndex, prepareTimelineFields, prepareTimelineLegendItems } from './utils';
-import { SeriesTableRow } from '@grafana/ui/src/components/VizTooltip';
+import { prepareTimelineFields, prepareTimelineLegendItems } from './utils';
+import { StateTimelineTooltip } from './StateTimelineTooltip';
 
 interface TimelinePanelProps extends PanelProps<TimelineOptions> {}
 
@@ -42,66 +34,23 @@ export const StateTimelinePanel: React.FC<TimelinePanelProps> = ({
   ]);
 
   const renderCustomTooltip = useCallback(
-    (alignedFrame: DataFrame, seriesIdx: number | null, datapointIdx: number | null) => {
+    (alignedData: DataFrame, seriesIdx: number | null, datapointIdx: number | null) => {
       // Not caring about multi mode in StateTimeline
-      if (seriesIdx === null && datapointIdx === null) {
+      if (seriesIdx === null || datapointIdx === null) {
         return null;
-      }
-      const xField = alignedFrame.fields[0];
-      const xFieldFmt = xField.display || getDisplayProcessor({ field: xField, timeZone, theme });
-
-      const field = alignedFrame.fields[seriesIdx!];
-      const dataFrameFieldIndex = field.state?.origin;
-      const fieldFmt = field.display || getDisplayProcessor({ field, timeZone, theme });
-      const value = field.values.get(datapointIdx!);
-      const display = fieldFmt(value);
-      const fieldDisplayName = dataFrameFieldIndex
-        ? getFieldDisplayName(
-            data.series[dataFrameFieldIndex.frameIndex].fields[dataFrameFieldIndex.fieldIndex],
-            data.series[dataFrameFieldIndex.frameIndex],
-            data.series
-          )
-        : null;
-
-      const nextStateIdx = findNextStateIndex(field, datapointIdx!);
-      let nextStateTs;
-      if (nextStateIdx) {
-        nextStateTs = xField.values.get(nextStateIdx!);
-      }
-
-      const stateTs = xField.values.get(datapointIdx!);
-
-      let toFragment = null;
-      let durationFragment = null;
-
-      if (nextStateTs) {
-        const duration =
-          nextStateTs && formattedValueToString(getValueFormat('dtdurationms')(nextStateTs - stateTs, 0));
-        durationFragment = (
-          <>
-            <br />
-            <strong>Duration:</strong> {duration}
-          </>
-        );
-        toFragment = (
-          <>
-            {' to'} <strong>{xFieldFmt(xField.values.get(nextStateIdx!)).text}</strong>
-          </>
-        );
       }
 
       return (
-        <div style={{ fontSize: theme.typography.bodySmall.fontSize }}>
-          {fieldDisplayName}
-          <br />
-          <SeriesTableRow label={display.text} color={display.color || FALLBACK_COLOR} isActive />
-          From <strong>{xFieldFmt(xField.values.get(datapointIdx!)).text}</strong>
-          {toFragment}
-          {durationFragment}
-        </div>
+        <StateTimelineTooltip
+          data={data.series}
+          alignedData={alignedData}
+          seriesIdx={seriesIdx}
+          datapointIdx={datapointIdx}
+          timeZone={timeZone}
+        />
       );
     },
-    [theme, timeZone, data]
+    [timeZone, data]
   );
 
   if (!frames || warn) {
@@ -123,7 +72,6 @@ export const StateTimelinePanel: React.FC<TimelinePanelProps> = ({
       height={height}
       legendItems={legendItems}
       {...options}
-      // hardcoded
       mode={TimelineMode.Changes}
     >
       {(config, alignedFrame) => {
